Separate execution script and main script class
