+++ slimdown-app.py
@@ -1,37 +1,9 @@
 #!/usr/bin/env python3
-import os
 from argparse import ArgumentParser
 
+from services.slimdown import Slimdown
 from helpers.filehelper import FileHelper
 from helpers.mdurlhelper import MarkdownUrlHelper
-from services.htmltomd import HtmlToMarkdown
-
-class Slimdown():
-    def __init__(self):
-        self.__markdownOutput = []
-
-    def __setMarkdownOutput(self, data):
-        self.__markdownOutput = data
-
-    def parse(self, htmlFile):
-        """Passes the HTML file to the Markdown parser. Gets the resulting md and stores it in its own property"""
-        htmlToMdParser = HtmlToMarkdown()
-        htmlToMdParser.readFile(htmlFile)
-        self.__setMarkdownOutput(htmlToMdParser.getMarkdown())
-    
-    def __writeToFile(self, mdFile):
-        """Write selfs markdown data to file"""
-        with open(mdFile, 'a') as writer:
-            for line in self.__markdownOutput: 
-                writer.write(line)
-
-    def outputMarkdown(self, mdFile):
-        """Public method to coordinate the writing out of markdown"""
-        if os.path.exists(mdFile):
-            os.remove(mdFile)
-
-        self.__writeToFile(mdFile)
-
 
 # Execution point 
 Slimdown = Slimdown()
